// packages/sdk/src/redux/actions.ts
export enum ReduxSdkActionTypes {
  SetInitialized = '@archanova-sdk/SetInitialized',
  SetConnected = '@archanova-sdk/SetConnected',
  SetAccount = '@archanova-sdk/SetAccount',
  SetAccountDevice = '@archanova-sdk/SetAccountDevice',
  SetDevice = '@archanova-sdk/SetDevice',
  SetEns = '@archanova-sdk/SetEns',
  SetEth = '@archanova-sdk/SetEth',
  SetIncomingAction = '@archanova-sdk/SetIncomingAction',
  SetAuthenticated = '@archanova-sdk/SetAuthenticated',
}